Filiere_Service/requirements.txt
Flask
Flask_SQLAlchemy
PyMySQL
python-dotenv
PyJWT
cryptography
flask-restx
requests